Extracted some code
